feat: logging response + content-encoding: gzip support
This PR makes three improvements

1. It adds logging response, a feature recently added in 8.6
2. It adds Content-Encoding: gzip support, a feature similarly
   recently added in 8.6.
3. It adds a workaround to support the .gz mime type
   (until https://github.com/Bogdanp/racket-net-mime-type/issues/1
   is fixed, that is)

For Racket <8.6, the first and second features won't be available.
We use version-case to accomplish this.

This PR should not be merged until
https://github.com/samth/version-case/pull/2 is first merged, however,
since right now there are a lot of unnecessary dependencies in
version-case. Once that's fixed, we can depend on it without worries.